+++ flutter_inappwebview_macos/macos/Classes/InAppWebView/InAppWebView.swift
@@ -48,6 +48,8 @@ public class InAppWebView: WKWebView, WKUIDelegate,
     
     var currentOpenPanel: NSOpenPanel?
     
+    fileprivate var interceptOnlyAsyncAjaxRequestsPluginScript: PluginScript?
+    
     init(id: Any?, plugin: InAppWebViewFlutterPlugin?, frame: CGRect, configuration: WKWebViewConfiguration,
          userScripts: [UserScript] = []) {
         super.init(frame: frame, configuration: configuration)
@@ -209,7 +211,11 @@ public class InAppWebView: WKWebView, WKUIDelegate,
         configuration.userContentController.addPluginScript(ORIGINAL_VIEWPORT_METATAG_CONTENT_JS_PLUGIN_SCRIPT)
         configuration.userContentController.addPluginScript(ON_SCROLL_CHANGED_EVENT_JS_PLUGIN_SCRIPT)
         if let settings = settings {
+            interceptOnlyAsyncAjaxRequestsPluginScript = createInterceptOnlyAsyncAjaxRequestsPluginScript(onlyAsync: settings.interceptOnlyAsyncAjaxRequests)
             if settings.useShouldInterceptAjaxRequest {
+                if let interceptOnlyAsyncAjaxRequestsPluginScript = interceptOnlyAsyncAjaxRequestsPluginScript {
+                    configuration.userContentController.addPluginScript(interceptOnlyAsyncAjaxRequestsPluginScript)
+                }
                 configuration.userContentController.addPluginScript(INTERCEPT_AJAX_REQUEST_JS_PLUGIN_SCRIPT)
             }
             if settings.useShouldInterceptFetchRequest {
@@ -598,7 +604,16 @@ public class InAppWebView: WKWebView, WKUIDelegate,
                                             enable: newSettings.useShouldInterceptAjaxRequest,
                                             pluginScript: INTERCEPT_AJAX_REQUEST_JS_PLUGIN_SCRIPT)
             } else {
-                newSettings.useShouldInterceptFetchRequest = false
+                newSettings.useShouldInterceptAjaxRequest = false
+            }
+        }
+        
+        if newSettingsMap["interceptOnlyAsyncAjaxRequests"] != nil && settings?.interceptOnlyAsyncAjaxRequests != newSettings.interceptOnlyAsyncAjaxRequests {
+            if let applePayAPIEnabled = settings?.applePayAPIEnabled, !applePayAPIEnabled,
+               let interceptOnlyAsyncAjaxRequestsPluginScript = interceptOnlyAsyncAjaxRequestsPluginScript {
+                enablePluginScriptAtRuntime(flagVariable: FLAG_VARIABLE_FOR_INTERCEPT_ONLY_ASYNC_AJAX_REQUESTS_JS_SOURCE,
+                                            enable: newSettings.interceptOnlyAsyncAjaxRequests,
+                                            pluginScript: interceptOnlyAsyncAjaxRequestsPluginScript)
             }
         }
         
@@ -2606,6 +2621,7 @@ if(window.\(JAVASCRIPT_BRIDGE_NAME)[\(_callHandlerID)] != null) {
             webMessageListener.dispose()
         }
         webMessageListeners.removeAll()
+        interceptOnlyAsyncAjaxRequestsPluginScript = nil
         if windowId == nil {
             configuration.userContentController.removeAllPluginScriptMessageHandlers()
             configuration.userContentController.removeScriptMessageHandler(forName: "onCallAsyncJavaScriptResultBelowIOS14Received")